Updated to Babel 6, based on trusktr/npm-package-skeleton
Initial global build (npm run build-global). Doesn't seem to work in a
static example, but works in a Meteor Blaze example. I'm not sure if
I've made an error or if there's something I don't know about the Custom
Element API. Even if I use the polyfill, the problem is the same: static
example doesn't load all the elements, while the Meteor example does,
despite that the markups are identical for both.