Added "don't update dist and docs in pull request"
Sharing feedback from @dmvaldman: Typically when contributing to an open-source project, the contributor only changes the source files as necessary, and it is up to the maintainer to build the bundles/documentation/etc upon a new release. This avoids a lot of merge conflicts, properly distributes credit, and is much better for testing and releasing.