Hurray for multiple-inheritance!
The Node and Scene calsses now extend ImperativeBase, which
extends TreeNode, but Node also extends Transformable (which extends
Sizeable) and Scene extends Sizeable, which is a subset of the
Transformable functionality, so that the scene can be easily sized but
the Scene doesn't need to be translated, rotated, etc.

Currently the multiple-inheritance is mixin-based, but I'd like to
finish the implementation I have at
https://gist.github.com/trusktr/05b9c763ac70d7086fe3a08c2c4fb4bf when
Proxy is available in Safari. It will be much nicer because classes will
not need to be written as mixins, just as plain classes. So, currently
the multiple inheritance looks like this:

```js
class Node extends Transformable.mixin(ImperativeBase) { ... }
```

where Transformable is being used as a class-factory-style mixin (see
http://justinfagnani.com/2015/12/21/real-mixins-with-javascript-classes
for more details on those). Once we switch to my multiple-inheritance
tool, it will be written like this:

```js
class Node extends multiple(Transformable, ImperativeBase) { ... }
```

and the classes will be written as plain ES6 classes instead of
functions that dynamically return a class.