infra: delete package-lock files as part of cleanup to ensure that next install does not have differing versions of dependencies when they shouldn't be
For example, a package-lock in three-projected-material ended up with the bootstrap process leaving it with a different version of Three.js than lume, although both packages specify a compatible range, and this caused a type error during lume build. Besides, we want to delete package locks to ensure that we have a higher chance of catching in-range breaking changes before Lume users.