update packages: namely they all now commit build outputs in dist using this repo's branch of lume/cli
- The whole lume repo (including submodules) will now be operable (just run a static web server in the repo) without having to run a build so that at the very least people can run all examples and can even submit change ideas in the plain JavaScript dist code for someone else to bring into the TypeScript source,
- plus we removed the global script build from Lume and hence no longer have to run Webpack but only TypeScript so that PowerShell users have a higher chance to succeed at building
- Because of the global script removal, all examples have been updated to import Lume using JavaScript module `import` syntax. Onto the future!

TLDR: We're trying to reduce as much friction possible from the dev
cycle, to make things easier to maintain. In an upcoming change we will
replace Karma with @web/test-runner to greatly simplify the test setup.
This will finally eliminate Webpack and the Karma dependencies which are
now getting old (and we all know things with many dependencies tend to
break after a while when sub-depency updates become incompatible, which
we don't want to have to deal with).

Basically, we're moving towards vanilla JavaScript modules everywhere to
simplify maintenance and focus on features. That'll be more exciting and
fun!