infra: update lume cli to run tests in the amazing `@web/test-runner` instead of in Karma and gain an waaaaaaay better test dev/debug experience with much higher speed, update dist based on lume cli's updated target:esnext mode, update tests.yml to call lume clean/build/test separately as they are no longer coupled (this alows the lume commands to be more efficient by not repeating any steps unnecessarily but requires just a little more control on the usage side).
Overall this greatly increases simplicity by removing a bunch of dependencies that were needed for Karma although it was still clunky with lots of issues, replacing the test setup with web test-runner which is built for vanilla ES Modules out of the box, and has none of the issues I had with Karma (f.e. source maps just work, headless windowing does not need to be configured as it is baked in, no more webpack configuration because tests are just vanilla ES Modules, test-runner config is very simple, and test-runner has amazing interactive debug options at the cli in watch mode that make it super easy to focus on tests and open them in your browser, and the speedup is just AWESOME).

I haven't even enabled visual regression testing for web test-runner in lume cli's case yet, but I've used it on other projects, and it is great. Lume is going to be better-tested while also being easier to maintain with low complexity.

I can't really express just how much better the new setup with web test-runner is. I highly recommend giving `@web/test-runner` a try in any web project for front-end unit tests!