fix: prevent `pointercancel` from happening and interfering with our interactions
To fix this, we set `touch-action: none` on `<lume-scene>` by default to disable default pointer behaviors. To be extra safe, we also call `event.preventDefault()` in pointer event handlers. We also capture the pointer on pointerdown so that dragging works well no matter if the pointer goes outside of the scene.