feat: add DOM events for model loading progress and errors, add DOM event for `sizechange`, fully deprecate the `Events` object and usage of `.on`.
Deprecations:

Using `.on()` for events is fully deprecated, and the method will be removed in a version after 0.3.0. Use `.addEventListener()` instead (or your favorite web framework's template event syntax).

- `el.on('sizechange', ({x, y, z}) => {})` becomes `el.addEventListener('sizechange', ({target: {calculatedSize: {x, y, z}}}) => {})` (or just read from the element inside the event handler instead of reading from `event.target`).
- `el.on('MODEL_LOAD', ({model}) => {})` becomes `el.addEventListener('load', ({target: {threeModel: model}}) => {})`
- `el.on('MODEL_ERROR', error => {})` becomes `el.addEventListener('error', ({error}) => {})`
- `el.on('PROGRESS', event => {})` becomes `el.addEventListener('progress', event => {})` (`event` is the same)

The `el.emit()` method is also deprecated to be removed in a version
after 0.3.0, use `el.dispatchEvent()` instead. Along with `.on()` and
`.emit()`, the events mentioned will be removed in a future release
after 0.3.0.