refactor: simplify `*Fling` classes by making them extend from `Settable`, update threejs minimum version requirement and fix some type errors, delete the `Scene.physicallyCorrectLights` property
BREAKING: `Scene.physicallyCorrectLights` has been removed, light mode
is always physically correct now, as it has been deprecated in Three.js
and is being removed. If you had this value as false (that was the
default) the lighting of your scene may appear darker now and will need
adjustment. Lighting uses physically-correct units (candela, based on
world units in meters) which mainly means that you'll typically need
higher values for light `intensity` values, for example, depending on
the size of your scene, try 100 or 1000 instead of 1. This is a nice
step ahead for more better and more realistic lighting by default.